Divide classes out into their own files
