+++ Assets/Scripts/Simulation/Adventures/AdventureBranchingComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c165d62a76dee3b40aaac2bfe76dad61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
