+++ Assets/Scripts/Simulation/Adventures/AdventureComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c608b3bbcaff54c499f8ecd9e79644e9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
