+++ Assets/Scripts/Simulation/Adventures/AdventureNarrationComponent.cs
@@ -1,0 +1,11 @@
+﻿using Sirenix.OdinInspector;
+using UnityEngine;
+
+namespace Game.Simulation
+{
+	public class AdventureNarrationComponent : AdventureComponent
+	{
+		[TextArea(15, 20), PropertyOrder(0), Title("Narration Text")]
+		public string text;
+	}
+}
