+++ Assets/Scripts/Simulation/Adventures/AdventureNarrationComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d20bb68c98c820147a488761cb3b9072
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
