+++ Assets/Scripts/Simulation/Adventures/AdventurePathComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5f91768b297a1ba43b79bdd9e0e5fc3c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
