+++ Assets/Scripts/Simulation/Adventures/AdventureSkillCheckComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4a4aaf0f13e46884893d72e293062b61
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
