+++ Assets/Scripts/Simulation/Adventures/AdventureTextComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e241c42e6089ae343b280d4b702520a1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
