+++ Assets/Scripts/Simulation/Adventures/IAdventureComponent.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d4cb92eb956511c4caa7ea29310edb09
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
